% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_helpers.R
\name{check_ip}
\alias{check_ip}
\title{Helper function to check current IP address}
\usage{
check_ip()
}
\value{
A character string with the current IP address.
}
\description{
\code{check_ip} queries whatismyipaddress.com to determine the current IP
address.
}
