% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_listings_kj.R
\name{download_listings_kj}
\alias{download_listings_kj}
\title{Helper function to download Craigslist or Kijiji listings}
\usage{
download_listings_kj(url_list, timeout = 1, proxies = NULL, quiet = FALSE)
}
\arguments{
\item{timeout, proxies, quiet}{TKTK}

\item{urls}{A character vector of URLs to be scraped.}
}
\value{
A list of HTML objects.
}
\description{
\code{download_listings)kj} scrapes listings from a list of URLs.
}
