% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_listings_kj.R
\name{get_listings_kj}
\alias{get_listings_kj}
\title{Helper function to download Craigslist or Kijiji listings}
\usage{
get_listings_kj(url_list, timeout = 1, proxies = NULL, quiet = FALSE)
}
\arguments{
\item{url_list}{A character vector of URLs to be scraped.}

\item{timeout, proxies, quiet}{TKTK}
}
\value{
A list of HTML objects.
}
\description{
\code{get_listings_kj} scrapes listings from a list of URLs.
}
