% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls_kj.R
\name{get_urls_kj}
\alias{get_urls_kj}
\title{Scrape Kijiji listing URLs}
\usage{
get_urls_kj(city_name, short_long, timeout = 1, proxies = NULL, quiet = FALSE)
}
\arguments{
\item{city_name}{A character string: the city to be scraped.}

\item{short_long}{A character string, either "short" or "long", to
determine whether STR or LTR listing URLs should be scraped.}

\item{proxies}{Character vector of IPs to use for proxy connections. If
the length is less than the number of processes set by `future::plan()`,
proxies will be recycled.}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A list of URLs.
}
\description{
\code{get_urls_kj} scrapes Kijiji listing URLs for a city.
}
