% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_bbox.R
\name{gg_bbox}
\alias{gg_bbox}
\title{Function to specify a bounding box for geom_sf}
\usage{
gg_bbox(geom, x1 = 0, x2 = 1, y1 = 0, y2 = 1, ...)
}
\arguments{
\item{geom}{TKTK}

\item{x1}{TKTK}

\item{x2}{TKTK}

\item{y1}{TKTK}

\item{y2}{TKTK}

\item{...}{TKTK}
}
\value{
TKTK.
}
\description{
\code{gg_bbox} TKTK.
}
\details{
A function for TKTK.
}
