% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_helpers.R
\name{\%do_upgo\%}
\alias{\%do_upgo\%}
\title{Helper function to execute parallel network code with proxies}
\usage{
obj \%do_upgo\% ex
}
\arguments{
\item{obj}{The sequence to iterate over}

\item{ex}{An expression to evaluate in parallel}
}
\value{
A list of evaluated expressions.
}
\description{
The function executes network code in a parallel future
with proxies
}
