% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_helpers.R
\name{handler_upgo}
\alias{handler_upgo}
\title{Helper function to set a progress reporting strategy}
\usage{
handler_upgo(message)
}
\arguments{
\item{message}{A character string describing the task being iterated.}
}
\value{
No visible output.
}
\description{
\code{handler_upgo} sets a progress reporting strategy.
}
