% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{helper_cl_urls}
\alias{helper_cl_urls}
\title{Helper function to scrape Craigslist listing URLs}
\usage{
helper_cl_urls(city_name)
}
\arguments{
\item{city_name}{A character scalar of the city to be scraped.}
}
\value{
A list of URLs.
}
\description{
\code{helper_cl_urls} scrapes Craigslist listing URLs for a city.
}
