% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kijiji_scrape_helpers.R
\name{helper_detail_parse}
\alias{helper_detail_parse}
\title{Helper function to generate Kijiji detail field}
\usage{
helper_detail_parse(.x)
}
\arguments{
\item{.x}{A single scraped Kijiji listing.}
}
\value{
A character scalar.
}
\description{
Helper function to generate Kijiji detail field
}
