% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_helpers.R
\name{helper_do_parallel}
\alias{helper_do_parallel}
\title{Helper function to execute paralle network code with proxies}
\usage{
helper_do_parallel(iterator, statement)
}
\arguments{
\item{iterator}{The sequence to iterate over}

\item{statement}{An expression to evaluate in parallel}
}
\value{
A list of evaluated expressions.
}
\description{
\code{helper_do_parallel} executes network code in a parallel future with
proxies
}
