% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{helper_download_listing}
\alias{helper_download_listing}
\title{Helper function to download Craigslist or Kijiji listings}
\usage{
helper_download_listing(url_list)
}
\arguments{
\item{url_list}{A character vector of URLs to be scraped.}
}
\value{
A list of HTML objects.
}
\description{
\code{helper_download_listing} scrapes listings from a list of URLs.
}
