% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{helper_error_cl}
\alias{helper_error_cl}
\title{Helper function to generate error Craigslist output}
\usage{
helper_error_cl(.y, city_name)
}
\arguments{
\item{.y}{A single Craigslist URL.}

\item{city_name}{A character string indicating the name of the city in which
the listing is located.}
}
\value{
A one-row data frame.
}
\description{
Helper function to generate error Craigslist output
}
