% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{helper_error_kj}
\alias{helper_error_kj}
\title{Helper function to generate error Kijiji output}
\usage{
helper_error_kj(.y, city_name)
}
\arguments{
\item{.y}{A single Kijiji URL.}

\item{city_name}{A character scalar indicating the name of the city in which
the listing is located.}
}
\value{
A one-row data frame.
}
\description{
Helper function to generate error Kijiji output
}
