% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ab_helpers.R
\name{helper_parse_ab}
\alias{helper_parse_ab}
\title{Helper function to parse location information from an Airbnb listing}
\usage{
helper_parse_ab(scrape_result)
}
\arguments{
\item{scrape_result}{A one-row data frame produced as an interim result from
\code{helper_scrape_ab}.}
}
\description{
\code{helper_parse_ab} parses the location (city, region and from a single
scraped Airbnb listing.
}
