% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cl_helpers.R
\name{helper_parse_cl}
\alias{helper_parse_cl}
\title{Helper function to parse scraped Craigslist listings}
\usage{
helper_parse_cl(.x, .y, city_name)
}
\arguments{
\item{.x}{A single scraped Craigslist listing, as retrieved with read_html.}

\item{.y}{A single Craigslist URL.}

\item{city_name}{A character scalar indicating the name of the city in which
the listing is located.}
}
\value{
A one-row data frame.
}
\description{
\code{helper_parse_cl} parses a scraped Craiglist listing.
}
