% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_helpers.R
\name{helper_require}
\alias{helper_require}
\title{Helper function to check for package requirement}
\usage{
helper_require(pkg_name)
}
\arguments{
\item{pkg_name}{A character string giving the name of a package.}
}
\value{
An error message if the package isn't found.
}
\description{
\code{helper_require} checks for a package and stops the function if the
package is not found.
}
