% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ab_helpers.R
\name{helper_scrape_ab}
\alias{helper_scrape_ab}
\title{Helper function to scrape location information from an Airbnb listing}
\usage{
helper_scrape_ab(PID)
}
\arguments{
\item{PID}{An Airbnb property ID to be scraped.}
}
\description{
\code{helper_scrape_ab} scrapes the location (city, region and
country) from a single Airbnb listing and forwards to the calling function
for further processing.
}
