% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{helper_scrape_ab_registration}
\alias{helper_scrape_ab_registration}
\title{Helper function to scrape registration information from an Airbnb listing}
\usage{
helper_scrape_ab_registration(PID)
}
\arguments{
\item{PID}{An Airbnb property ID to be scraped.}
}
\description{
\code{helper_scrape_ab_registration} scrapes the registration number from a
single Airbnb listing and forwards to the calling function for further
processing.
}
