% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{helper_scrape_listing_page_kj}
\alias{helper_scrape_listing_page_kj}
\title{Helper function to scrape listing page for Kijiji}
\usage{
helper_scrape_listing_page_kj(url, user_agent, proxy)
}
\arguments{
\item{url, user_agent, proxy}{Arguments passed on from the main function.}
}
\description{
Helper function to scrape listing page for Kijiji
}
