% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scrape_location.R
\name{helper_scrape_location_parse}
\alias{helper_scrape_location_parse}
\title{Helper function to scrape location information from an Airbnb listing}
\usage{
helper_scrape_location_parse(results)
}
\arguments{
\item{results}{The interim results from helper_scrape_location.}
}
\description{
\code{helper_scrape_location_parse} scrapes the location (city, region and
country) from a single Airbnb listing and forwards to the calling function
for further processing.
}
