% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scrape_location_parse_2.R
\name{helper_scrape_location_parse_2}
\alias{helper_scrape_location_parse_2}
\title{Helper function to scrape location information from an Airbnb listing}
\usage{
helper_scrape_location_parse_2(scrape_result)
}
\arguments{
\item{scrape_result}{A one-row data frame produced as an interim result from
\code{helper_scrape_location}.}
}
\description{
\code{helper_scrape_location_parse} scrapes the location (city, region and
country) from a single Airbnb listing and forwards to the calling function
for further processing.
}
