% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{helper_urls_kj}
\alias{helper_urls_kj}
\title{Helper function to scrape Kijiji listing URLs}
\usage{
helper_urls_kj(city_name, short_long)
}
\arguments{
\item{city_name}{A character string: the city to be scraped.}

\item{short_long}{A character string, either "short" or "long", to determine
whether STR or LTR listing URLs should be scraped.}
}
\value{
A list of URLs.
}
\description{
\code{helper_urls_kj} scrapes Kijiji listing URLs for a city.
}
