% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_listing_kj.R
\name{parse_listing_kj}
\alias{parse_listing_kj}
\title{Helper function to parse scraped Kijiji listings}
\usage{
parse_listing_kj(x, city_name, proxies = NULL, quiet = FALSE)
}
\arguments{
\item{x}{A single scraped Kijiji listing, as retrieved with
\code{get_listings_kj}.}

\item{city_name}{A character scalar indicating the name of the city in which
the listing is located.}
}
\value{
A one-row data frame.
}
\description{
\code{helper_parse_kj} parses a scraped Kijiji listing.
}
