% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_matchr.R
\name{reinstall_matchr}
\alias{reinstall_matchr}
\title{Function to reinstall the matchr package from GitHub}
\usage{
reinstall_matchr(branch = NULL)
}
\arguments{
\item{branch}{A character string identifying the branch to be installed, if
not the master branch (default).}
}
\value{
There is no tangible output to the function; it simply detaches and
updates the matchr package.
}
\description{
Function to reinstall the matchr package from GitHub
}
