% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_strr.R
\name{reinstall_strr}
\alias{reinstall_strr}
\title{Function to reinstall the strr package from GitHub}
\usage{
reinstall_strr(branch = NULL)
}
\arguments{
\item{branch}{A character string identifying the branch to be installed, if
not the master branch (default).}
}
\value{
There is no tangible output to the function, simply the detaching,
updating and reattaching of the strr package.
}
\description{
Function to reinstall the strr package from GitHub
}
