% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_upgo.R
\name{reinstall_upgo}
\alias{reinstall_upgo}
\title{Function to reinstall the upgo package from GitHub}
\usage{
reinstall_upgo(branch = NULL)
}
\arguments{
\item{branch}{A character string identifying the branch to be installed, if
not the master branch (default).}
}
\value{
There is no tangible output to the function, simply the detaching and
updating the upgo package.
}
\description{
Function to reinstall the upgo package from GitHub
}
