% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_aws_download.R
\name{upgo_aws_download}
\alias{upgo_aws_download}
\title{Function to download file from the UPGo AWS bucket}
\usage{
upgo_aws_download(object_name, file_path = NULL, quiet = FALSE)
}
\arguments{
\item{object_name}{A character string which identifies the file name to be
downloaded.}

\item{file_path}{A character string which optionally identifies a subfolder
of the working directory in which to save the output.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
The output will be a file saved to disk in the current working
directory (possibly modified by the `file_path` argument).
}
\description{
\code{upgo_aws_download} downloads a specified file from the UPGo AWS bucket
}
\details{
This is a convenience wrapper around \code{aws.s3::save_object} to download
a single object from the UPGo AWS bucket. It will only work with appropriate
values for `AWS_ACCESS_KEY_ID` and `AWS_SECRET_ACCESS_KEY` saved in
.Renviron.
}
