% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_disconnect.R
\name{upgo_disconnect}
\alias{upgo_disconnect}
\title{Function to connect to UPGo server}
\usage{
upgo_disconnect()
}
\value{
The function returns no output, but removes objects from the global
environment.
}
\description{
\code{upgo_disconnect} closes an open connection to the UPGo server and
removes the associated tables from the global environment.
}
\details{
A function for disconnecting from the UPGo server and any associated tables.
An open connection with the name `con` will be removed from the special upgo
package environment (`.upgo_env`), as will tables named `property_remote`,
`daily_remote`, `daily_inactive_remote`, `host_remote`,
`host_inactive_remote`, `review_remote`, `review_user_remote`,
`review_text_remote`, `geolocation_remote` and `ha_mapping_remote`.
}
