% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_helpers.R
\name{upgo_do}
\alias{upgo_do}
\title{Aliases to foreach functions}
\usage{
upgo_do(obj, ex)
}
\arguments{
\item{obj}{The `foreach` statement.}

\item{ex}{The expression to be iterated.}
}
\description{
Aliases to foreach functions
}
