% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_download.R
\name{upgo_download}
\alias{upgo_download}
\title{Download URLs with progress reporting}
\usage{
upgo_download(urls, quiet = FALSE)
}
\arguments{
\item{urls}{A character vector of URLs to download.}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A list of HTML objects.
}
\description{
\code{upgo_download} is a thin wrapper around helper_download_listing to
enable progress reporting when used outside the main UPGo scraping functions.
}
