% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_helpers.R
\docType{data}
\name{upgo_env}
\alias{upgo_env}
\title{UPGo package environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
upgo_env
}
\description{
UPGo package environment
}
\keyword{datasets}
