% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_ab.R
\name{upgo_scrape_ab}
\alias{upgo_scrape_ab}
\title{Scrape location information from Airbnb listings}
\usage{
upgo_scrape_ab(property, quiet = FALSE)
}
\arguments{
\item{property}{An input table with a field named `property_ID` which will be
used to generate URLs for scraping.}

\item{quiet}{A logical scalar Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A table with property_ID, city, region, and country, along with the
raw strings used to identify the location and the date on which the scrape
was performed.
}
\description{
\code{upgo_scrape_ab} scrapes location (city, region and country) from
Airbnb listings.
}
\details{
TKTK
}
