% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_ab_registration.R
\name{upgo_scrape_ab_registration}
\alias{upgo_scrape_ab_registration}
\title{Scrape registration numbers from Airbnb listings}
\usage{
upgo_scrape_ab_registration(property, timeout = 1, quiet = FALSE)
}
\arguments{
\item{property}{An input table with a field named `property_ID` which will be
used to generate URLs for scraping.}

\item{timeout}{TKTK}

\item{quiet}{A logical scalar Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A table with property_ID, registration and date fields.
}
\description{
\code{upgo_scrape_ab_registration} scrapes registration numbers from
Airbnb listings.
}
\details{
TKTK
}
