% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_cl.R
\name{upgo_scrape_cl}
\alias{upgo_scrape_cl}
\title{Function to scrape Craigslist listings}
\usage{
upgo_scrape_cl(
  city,
  old_results = NULL,
  recovery = FALSE,
  proxies = NULL,
  quiet = FALSE
)
}
\arguments{
\item{city}{A character scalar indicating the city to scrape. It must
correspond to the domain prefix used for a city's Craigslist site (e.g.
"sfbay" for the San Francisco Bay area).}

\item{old_results}{A data frame. If the output of a previous run of
\code{upgo_scrape_cl} is supplied, listings previously scraped will
be incorporated into the new results.}

\item{recovery}{A logical vector. Should the function attempt to recover
results from a previous, unsuccessful function call?}

\item{proxies}{Character vector of IPs to use for proxy connections.}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A table with one row per listing scraped.
}
\description{
\code{upgo_scrape_cl} scrapes all data from Craigslist rental
listings in a given city.
}
