% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_craigslist.R
\name{upgo_scrape_craigslist}
\alias{upgo_scrape_craigslist}
\title{Function to scrape Craigslist listings}
\usage{
upgo_scrape_craigslist(city, old_results = NULL, timeout = 0.2, quiet = FALSE)
}
\arguments{
\item{city}{A character scalar indicating the city to scrape. It must
correspond to the domain prefix used for a city's Craigslist site (e.g.
"sfbay" for the San Francisco Bay area).}

\item{old_results}{A data frame. If the output of a previous run of
\code{upgo_scrape_craigslist} is supplied, listings previously scraped will
not be scraped again.}

\item{timeout}{A positive numeric scalar. How long in seconds should the
function pause between scrape attempts to avoid triggering cooldowns?}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A table with one row per listing scraped.
}
\description{
\code{upgo_scrape_craigslist} scrapes all data from Craigslist rental
listings in a given city.
}
