% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_connect.R
\name{upgo_scrape_disconnect}
\alias{upgo_scrape_disconnect}
\title{Function to close Selenium server after completing webscraping}
\usage{
upgo_scrape_disconnect()
}
\value{
The function returns no output, but closes an open Selenium server
and performs associated cleanup.
}
\description{
\code{upgo_scrape_disconnect} closes a Selenium server and removes the
connection object from the global environment.
}
\details{
This function closes any active Selenium server and any web driver currently
assigned to the object `remDr`, removes both objects, and performs garbage
collection on memory, to ensure #' that subsequent scraping connections will
be made successfully.
}
