% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_kijiji.R
\name{upgo_scrape_kijiji}
\alias{upgo_scrape_kijiji}
\title{Function to scrape Kijiji listings}
\usage{
upgo_scrape_kijiji(
  city,
  old_results = NULL,
  short_long = "both",
  timeout = 0.2,
  quiet = FALSE
)
}
\arguments{
\item{city}{A character scalar indicating the city to scrape. Currently
accepts "montreal", "toronto" or "vancouver" as inputs.}

\item{old_results}{A data frame. If the output of a previous run of
\code{upgo_scrape_kijiji} is supplied, listings previously scraped will not
be scraped again.}

\item{short_long}{A character scalar. Should short-term rentals ("short"),
long-term rentals ("long") or both ("both", the default) be scraped?}

\item{timeout}{A positive numeric scalar. How long in seconds should the
function pause between scrape attempts to avoid triggering cooldowns?}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A table with one row per listing scraped.
}
\description{
\code{upgo_scrape_kijiji} scrapes all data from Kijiji rental listings in
Montreal, Toronto or Vancouver.
}
