% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_location.R
\name{upgo_scrape_location}
\alias{upgo_scrape_location}
\title{Function to scrape location information from Airbnb listings}
\usage{
upgo_scrape_location(
  property,
  port = 4445L,
  chunk_size = 100,
  cores = 1L,
  quiet = FALSE
)
}
\arguments{
\item{property}{An input table with a field named \code{property_ID} which
will be used to generate URLs for scraping.}

\item{port}{A positive integer scalar. What port is the Selenium server
running on? Set this to match the value passed through
[func(upgo_scrape_connect)].}

\item{chunk_size}{The number of listings to scrape in a single batch.
Higher numbers will be slightly faster, but provide less frequent progress
feedback and higher risk of lost data if the scraper encounters an error.}

\item{cores}{A positive integer scalar. How many processing cores should be
used to perform the computationally intensive intersection steps? The
implementation of multicore processing does not support Windows, so this
argument should be left with its default value of 1 in those cases.}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\description{
\code{upgo_scrape_location} scrapes location (city, region and country) from
Airbnb listings.
}
\details{
TKTK
}
