% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_location.R
\name{upgo_scrape_location_helper}
\alias{upgo_scrape_location_helper}
\title{Helper function to scrape location information from an Airbnb listing}
\usage{
upgo_scrape_location_helper(PID)
}
\arguments{
\item{PID}{An Airbnb property ID to be scraped.}
}
\description{
\code{upgo_scrape_location_helper} scrapes the location (city, region and
country) from a single Airbnb listing.
}
\details{
TKTK
}
