% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_nu.R
\name{upgo_scrape_nu}
\alias{upgo_scrape_nu}
\title{Scrape media articles from Nexis Uni}
\usage{
upgo_scrape_nu(
  query,
  start_date = NULL,
  end_date = NULL,
  language = "English",
  skip = 0,
  file_prefix = substr(query, 1, 6),
  timeout = 20,
  quiet = FALSE
)
}
\arguments{
\item{query}{TKTK}

\item{start_date}{TKTK}

\item{end_date}{TKTK}

\item{language}{TKTK}

\item{skip}{A positive integer. How many batches of 50 articles have already
been downloaded, and so should be skipped? E.g. a value of 33 means that 33
files (of 50 articles each) have been downloaded, for a total of 1,650
articles, and the function will resume with article 1,651.}

\item{file_prefix}{TKTK}

\item{timeout}{TKTK}

\item{quiet}{A logical scalar Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
TKTK
}
\description{
\code{upgo_scrape_ln} scrapes Nexis Uni for all articles which mention a
given search term.
}
\details{
TKTK
}
