% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kijiji_scrape_helpers.R
\name{helper_cl_urls}
\alias{helper_cl_urls}
\title{Helper function to scrape Craigslist listing URLs}
\usage{
helper_cl_urls(city_name, quiet)
}
\arguments{
\item{city_name}{A character scalar of the city to be scraped.}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function?}
}
\value{
A list of URLs.
}
\description{
\code{helper_cl_urls} scrapes Craigslist listing URLs for a city.
}
