% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kijiji_scrape_helpers.R
\name{helper_download_listing}
\alias{helper_download_listing}
\title{Helper function to download Craigslist or Kijiji listings}
\usage{
helper_download_listing(url_list, url_prefix, url_suffix, quiet)
}
\arguments{
\item{url_list}{A character vector of URLs to be scraped.}

\item{url_prefix}{A character scalar to be prepended to each URL.}

\item{url_suffix}{A character scalar to be appended to each URL.}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function?}
}
\value{
A list of HTML objects.
}
\description{
\code{helper_download_listing} scrapes listings from a list of URLs.
}
