% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{helper_error_cl}
\alias{helper_error_cl}
\title{Helper function to generate error Craigslist output}
\usage{
helper_error_cl()
}
\value{
A zero-row data frame.
}
\description{
Helper function to generate error Craigslist output
}
