% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{helper_error_kj}
\alias{helper_error_kj}
\title{Helper function to generate error Kijiji output}
\usage{
helper_error_kj()
}
\value{
A zero-row data frame.
}
\description{
Helper function to generate error Kijiji output
}
