% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_intersect_summarize.R
\name{st_intersect_summarize}
\alias{st_intersect_summarize}
\title{Function to perform areal interpolation of variables}
\usage{
st_intersect_summarize(data, destination, group_vars = NULL, population,
  count_vars, mean_vars)
}
\arguments{
\item{data}{TKTK}

\item{destination}{TKTK}

\item{group_vars}{TKTK}

\item{population}{TKTK}

\item{count_vars}{TKTK}

\item{mean_vars}{TKTK}
}
\value{
The function returns a table with the variables of interest
reassembled in the `destination` geometry (potentially modified by
`group_vars`).
}
\description{
\code{st_intersect_summarize} performs areal interpolation of variables from
one geometry to another.
}
\details{
A function for translating variables from one polygon geometry to another,
using constant areal interpolation. The function works with variables which
represent counts within a polygon (`count_vars`) and with variables which
represent averages across a polygon (`mean_vars`). Using an additional
variable representing the population (or other weighting factor) of each
polygon (`population`), the function reapportions the variables from the
origin geometry (in `data`) to the destination geometry (`destination`),
using an optional `group_vars` argument to control how the destination
geometry is assembled. It works by assuming that variables are evenly
distributed throughout the polygons.
}
