% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_connect.R
\name{upgo_connect}
\alias{upgo_connect}
\title{Function to connect to UPGo server}
\usage{
upgo_connect(property = TRUE, daily = TRUE, multi = TRUE, reviews = FALSE)
}
\arguments{
\item{property}{A logical scalar. Should the property table (named
`property_all`) be added to the global environment?}

\item{daily}{A logical scalar. Should the daily table (named `daily_all`) be
added to the global environment?}

\item{multi}{A logical scalar. Should the multilisting table (named
`multi_all`) be added to the global environment?}

\item{reviews}{A logical scalar. Should the reviews table (named
`reviews_all`) be added to the global environment?}
}
\value{
The function returns no output, but makes assignments to the global
environment.
}
\description{
\code{upgo_connect} opens a connection to the UPGo server and optionally adds
tables from the server to the global environment.
}
\details{
A function for connecting to the UPGo server. For the function to work, the
variables `PGHOST`, `PGUSER` and `PGPASSWORD` need to be defined in the
user's .Renviron file. A connection with the name `con` will be added to the
global environment. The function will also optionally add tables from the
server to the global environment.
}
