% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_connect.R
\name{upgo_connect}
\alias{upgo_connect}
\title{Function to connect to UPGo server}
\usage{
upgo_connect(
  property = TRUE,
  daily = TRUE,
  daily_inactive = FALSE,
  host = TRUE,
  host_inactive = FALSE,
  review = FALSE,
  review_user = FALSE,
  review_text = FALSE,
  geolocation = FALSE,
  remote = FALSE
)
}
\arguments{
\item{property}{A logical scalar. Should the property table (named
`property_all`) be added to the global environment?}

\item{daily}{A logical scalar. Should the daily table (named `daily_all`) be
added to the global environment?}

\item{daily_inactive}{A logical scalar. Should the daily_inactive table
(named `daily_inactive_all`) be added to the global environment?}

\item{host}{A logical scalar. Should the host table (named `host_all`) be
added to the global environment?}

\item{host_inactive}{A logical scalar. Should the host_inactive table (named
`host_inactive_all`) be added to the global environment?}

\item{review}{A logical scalar. Should the review table (named
`review_all`) be added to the global environment?}

\item{review_user}{A logical scalar. Should the review_user table (named
`review_user_all`) be added to the global environment?}

\item{review_text}{A logical scalar. Should the review_text table (named
`review_text_all`) be added to the global environment?}

\item{geolocation}{A logical scalar. Should the geolocation table (named
`geolocation_remote`) be added to the global environment?}

\item{remote}{A logical scalar. Should a remote connection be opened even if
only local tables are being connected?}
}
\value{
The function returns no output, but makes assignments to the global
environment.
}
\description{
\code{upgo_connect} opens a connection to the UPGo server and optionally adds
tables from the server to the global environment.
}
\details{
A function for connecting to the UPGo server. For the function to work, the
variables `PGHOST`, `PGUSER` and `PGPASSWORD` need to be defined in the
user's .Renviron file. A connection with the name `con` will be added to the
global environment. The function will also optionally add tables from the
server to the global environment.
}
