% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_location_scrape.R
\name{upgo_location_scrape}
\alias{upgo_location_scrape}
\title{Function to scrape location information from Airbnb listings}
\usage{
upgo_location_scrape(property, delay = 10, geography = NULL)
}
\arguments{
\item{property}{An input table with a field named \code{property_ID} which
will be used to generate URLs for scraping.}

\item{delay}{A numeric scalar. The number of seconds the function should
delay in between listings to prevent connection errors or Airbnb shutting
down access.}

\item{geography}{An existing output table which the function should add to.
By default the function creates a new table instead.}
}
\description{
\code{upgo_location_scrape} scrapes location (city, region and country) from
Airbnb listings.
}
\details{
TKTK
}
