% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_ab_registration.R
\name{upgo_scrape_ab_registration}
\alias{upgo_scrape_ab_registration}
\title{Scrape registration numbers from Airbnb listings}
\usage{
upgo_scrape_ab_registration(
  property,
  proxies = NULL,
  cores = 1L,
  quiet = FALSE
)
}
\arguments{
\item{property}{An input table with a field named `property_ID` which will be
used to generate URLs for scraping.}

\item{proxies}{Character vector of IPs to use for proxy connections. If
supplied, this must be at least as long as the number of cores.}

\item{cores}{A positive integer scalar. How many processing cores should be
used to scrape?}

\item{quiet}{A logical scalar Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A table with property_ID, registration and date fields.
}
\description{
\code{upgo_scrape_ab_registration} scrapes registration numbers from
Airbnb listings.
}
\details{
TKTK
}
