% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_connect.R
\name{upgo_scrape_connect}
\alias{upgo_scrape_connect}
\title{Function to open Selenium server for webscraping}
\usage{
upgo_scrape_connect(port = 4445L)
}
\arguments{
\item{port}{A positive integer scalar. What port should the Selenium server
run on?}
}
\value{
The function returns a connection object, which should be assigned to
`rD`.
}
\description{
\code{upgo_scrape_connect} opens a Selenium server to allow webscraping.
}
\details{
TKTK
}
