% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_connect.R
\name{upgo_scrape_connect}
\alias{upgo_scrape_connect}
\title{Function to open Selenium server for webscraping}
\usage{
upgo_scrape_connect(chrome = "83.0.4103.39", proxy = NULL, ...)
}
\arguments{
\item{chrome}{A character string specifying the version of Chrome to be used
with Selenium.}

\item{proxy}{A character string. The URL of a proxy server to connect
through.}

\item{...}{Additional arguments passed to RSelenium::rsDriver.}
}
\value{
The function returns a connection object, which it assigns to `rD` in
the global environment.
}
\description{
\code{upgo_scrape_connect} opens a Selenium server to allow webscraping.
}
\details{
This function opens a Selenium server to prepare for webscraping using one of
the upgo_scrape_* functions. For compatibility with those functions, it
assigns an object `rD` to the global environment, and will overwrite any
existing object with the same name without warning.
}
