% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_connect.R
\name{upgo_scrape_disconnect}
\alias{upgo_scrape_disconnect}
\title{Function to close Selenium server after completing webscraping}
\usage{
upgo_scrape_disconnect(connection = rD)
}
\arguments{
\item{connection}{A symbol corresponding to the open connection object.}
}
\value{
The function returns no output, but closes an open Selenium server
and performs associated cleanup.
}
\description{
\code{upgo_scrape_disconnect} closes a Selenium server and removes the
connection object from the global environment.
}
\details{
TKTK
}
