% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgo_scrape_location.R
\name{upgo_scrape_location}
\alias{upgo_scrape_location}
\title{Function to scrape location information from Airbnb listings}
\usage{
upgo_scrape_location(property, chunk_size = 100, cores = 1L, quiet = FALSE)
}
\arguments{
\item{property}{An input table with a field named `property_ID` which will be
used to generate URLs for scraping.}

\item{chunk_size}{A positive integer. The number of listings to be scraped in
one batch. Higher values will increase performance, but may be more
vulnerable to rate-limiting and data loss.}

\item{cores}{A positive integer scalar. How many processing cores should be
used to perform the computationally intensive intersection steps? The
implementation of multicore processing does not support Windows, so this
argument should be left with its default value of 1 in those cases.}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A table with property_ID, city, region, and country, along with the
raw strings used to identify the location and the date on which the scrape
was performed.
}
\description{
\code{upgo_scrape_location} scrapes location (city, region and country) from
Airbnb listings.
}
\details{
TKTK
}
